pyproject.toml
[tool.poetry]
name = "lab13-2021-base"
version = "0.1.0"
description = ""
authors = ["Roman Bartusiak <roman@stashaway.com>"]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.43.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
grpcio-tools = "^1.43.0"
poethepoet = "0.12.0"

[tool.poe.tasks]
grpc="python -m grpc_tools.protoc -I. --python_out=lab13_2021_base --grpc_python_out=lab13_2021_base protobuf/sample.proto"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
